long roomId로 변경, 210, 211 프로토콜 추가

+++ NetworkOmok_Client/src/OmokRoom.java
@@ -2,12 +2,12 @@ import java.util.Vector;
 
 public class OmokRoom {
 	public String roomName;
-	public int roomId;
+	public long roomId;
 	public int peopleCount = 0;
 	public Vector<String> player = new Vector<>();
 	public Vector<String> viewer = new Vector<>();
 	
-	public OmokRoom(int roomId, String creater) {
+	public OmokRoom(long roomId, String creater) {
 		this.roomId = roomId;
 		player.add(creater);
 	}
